refactor(product list): move noindex exemptions
This is a direct copy paste of the exemptions into a new file.
