+++ frontend/seo/product-list/index.ts
@@ -1,0 +1,1 @@
+export * from './noIndexExemptions';
